implementing crud operations on product

+++ src/main/java/com/wisan/project/services/CategoryService.java
@@ -3,10 +3,14 @@ package com.wisan.project.services;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.wisan.project.entities.Category;
 import com.wisan.project.repositories.CategoryRepository;
+import com.wisan.project.service.exceptions.ResourceNotFundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class CategoryService {
@@ -19,7 +23,7 @@ public class CategoryService {
 	}
 
 	public Category findById(Long id) {
-		return categoryRepository.findById(id).get();
+		return categoryRepository.findById(id).orElseThrow(() -> new ResourceNotFundException(id));
 	}
 
 	public Category insert(Category category) {
@@ -27,18 +31,31 @@ public class CategoryService {
 	}
 
 	public Category update(Long id, Category obj) {
-		Category entity = categoryRepository.getReferenceById(id);
-		updateData(entity, obj);
-		return categoryRepository.saveAndFlush(entity);
+		try {
+			Category entity = categoryRepository.getReferenceById(id);
+			updateData(entity, obj);
+			return categoryRepository.saveAndFlush(entity);
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFundException(id);
+		}
 	}
 
 	private void updateData(Category entity, Category obj) {
 		entity.setName(obj.getName());
 
 	}
+	
 
 	public void delete(Long id) {
-		categoryRepository.deleteById(id);
+		try {
+			if (categoryRepository.existsById(id)) {
+				categoryRepository.deleteById(id);
+			} else {
+				throw new ResourceNotFundException(id);
+			}
+		} catch (DataIntegrityViolationException e) {
+			throw new DataIntegrityViolationException("Data integrity violation occurred.");
+		}
 	}
 
 }
